+++ CHANGELOG.rst
@@ -8,6 +8,8 @@ Change log
 - Made our inline upload JavaScript specify its dependency on
   ``django.jQuery``.
 - Verified compatibility with Django 3.0.
+- Changed the abstract file model to protect files against cascading
+  folder deletions.
 
 
 `0.9`_ (2019-04-15)